# WIP: improve speed of coordinate transformations for arrays of obstime