--- a/astropy/coordinates/astrom_manager.py
+++ b/astropy/coordinates/astrom_manager.py
@@ -40,6 +40,7 @@ class transform_precision(ScienceState):
 
 def get_astrom(frame, tcode, precision=None):
     """
+    TBD
     """
 
     assert tcode in ['apio13', 'apci', 'apcs', 'apci13', 'apco13']
@@ -56,32 +57,45 @@ def get_astrom(frame, tcode, precision=None):
         mjd_resolution = precision / 86400.  # in days
 
     obstime = frame.obstime
+    jd1_tt, jd2_tt = get_jd12(obstime, 'tt')
 
     if mjd_resolution:
-        # apply binning to obstime
-        mjd_binned = np.int64(obstime.mjd / mjd_resolution + 0.5)
-        mjd_u, mjd_idx, mjd_uidx = np.unique(
-            mjd_binned, return_index=True, return_inverse=True
-            )
+        # compute mjd support points for interpolation of Earth pv and cip
+        mjd_lower = np.int64(obstime.mjd / mjd_resolution)
+        mjd_upper = mjd_lower + 1
+        mjd_u = np.unique([mjd_lower, mjd_upper])  # does sorting
 
-        # obstime_binned = obstime[mjd_idx]
-        obstime_binned = Time(mjd_u * mjd_resolution, format='mjd', scale=obstime.scale)
-    else:
-        obstime_binned = obstime
+        obstime_support = Time(mjd_u * mjd_resolution, format='mjd', scale=obstime.scale)
 
     if tcode in ['apci', 'apcs']:
         # find the position and velocity of earth
-        jd1_tt, jd2_tt = get_jd12(obstime_binned, 'tt')
-        earth_pv, earth_heliocentric = prepare_earth_position_vel(obstime_binned)
+        if mjd_resolution:
+            (
+                earth_pv_support, earth_heliocentric_support
+                ) = prepare_earth_position_vel(obstime_support)
+            # do interpolation
+            earth_pv = np.empty((len(obstime), 2, 3))
+            earth_heliocentric = np.empty((len(obstime), 3))
+            for dim in range(3):
+                for ipv in range(2):
+                    earth_pv[:, ipv, dim] = np.interp(
+                        obstime.mjd, obstime_support.mjd, earth_pv_support[:, ipv, dim]
+                        )
+                earth_heliocentric[:, dim] = np.interp(
+                        obstime.mjd, obstime_support.mjd, earth_heliocentric_support[:, dim]
+                        )
+
+        else:
+            earth_pv, earth_heliocentric = prepare_earth_position_vel(obstime)
 
     if tcode == 'apio13':
 
         lon, lat, height = frame.location.to_geodetic('WGS84')
 
-        xp, yp = get_polar_motion(obstime_binned)
-        jd1_utc, jd2_utc = get_jd12(obstime_binned, 'utc')
-        dut1utc = get_dut1utc(obstime_binned)
-        astrom_binned = erfa.apio13(
+        xp, yp = get_polar_motion(obstime)
+        jd1_utc, jd2_utc = get_jd12(obstime, 'utc')
+        dut1utc = get_dut1utc(obstime)
+        astrom = erfa.apio13(
             jd1_utc, jd2_utc, dut1utc,
             lon.to(u.radian).value, lat.to(u.radian).value,
             height.to(u.m).value,
@@ -95,8 +109,19 @@ def get_astrom(frame, tcode, precision=None):
 
     elif tcode == 'apci':
 
-        x, y, s = get_cip(jd1_tt, jd2_tt)
-        astrom_binned = erfa.apci(jd1_tt, jd2_tt, earth_pv, earth_heliocentric, x, y, s)
+        if mjd_resolution:
+            jd1_tt_support, jd2_tt_support = get_jd12(obstime_support, 'tt')
+            cip_support = get_cip(jd1_tt_support, jd2_tt_support)
+            # do interpolation; cip is an ordinary array
+            cip = tuple(
+                np.interp(obstime.mjd, obstime_support.mjd, cip_support[i])
+                for i in range(3)
+                )
+
+        else:
+            cip = get_cip(jd1_tt, jd2_tt)
+
+        astrom = erfa.apci(jd1_tt, jd2_tt, earth_pv, earth_heliocentric, *cip)
 
     elif tcode == 'apcs':
 
@@ -107,19 +132,11 @@ def get_astrom(frame, tcode, precision=None):
                 (frame.obsgeoloc.get_xyz(xyz_axis=-1).value[..., np.newaxis, :],
                  frame.obsgeovel.get_xyz(xyz_axis=-1).value[..., np.newaxis, :]),
                 axis=-2)
-        astrom_binned = erfa.apcs(jd1_tt, jd2_tt, pv, earth_pv, earth_heliocentric)
+        astrom = erfa.apcs(jd1_tt, jd2_tt, pv, earth_pv, earth_heliocentric)
 
     elif tcode == 'apci13':
         pass
     elif tcode == 'apco13':
         pass
 
-    if mjd_resolution:
-        astrom = astrom_binned[mjd_uidx]
-        jd1_ut1, jd2_ut1 = get_jd12(obstime, 'ut1')
-        astrom = erfa.aper13(jd1_ut1, jd2_ut1, astrom)
-    else:
-        astrom = astrom_binned
-
     return astrom
-