The erfa.astrom struct needs a substantial amount of computing power to update Earth parameters. However, all but the Earth rotation angle are relatively slowly changing numbers. If some accuracy can be sacrificed, calculating the slowly changing astrom parameters on certain time-steps only, can lead to large speed-ups. For example, a time resolution of 10 minutes will still have an accuracy of <~50 mas if ICRS is transformed to AltAz.

The attached commit is merely a proof of concept. Since this is my first PR, I'd appreciate some guidance on how to proceed. Below, I add some additional information, ideas and questions that should be discussed before I'll work on a final solution.

## The Problem
At our observatory, we often observe in AltAz frame for a session and then need to convert the coordinates to ICRS. A minimal example:
```python
import numpy as np
from astropy import coordinates
from astropy.coordinates import SkyCoord  # High-level coordinates
from astropy.coordinates import ICRS, CIRS, AltAz, EarthLocation
from astropy.coordinates import Angle, Latitude, Longitude  # Angles
from astropy.time import Time
from astropy import _erfa as erfa
import astropy.units as u
from astropy.utils.misc import NumpyRNGContext

numcoords = 10 ** 5

with NumpyRNGContext(1):
    # in reality, the following quantities are of course not random
    _ra = np.random.uniform(-179, 179, numcoords)
    _dec = np.random.uniform(-89, 89, numcoords)
    _time_delta = np.random.uniform(0, 1, numcoords)

ra = Longitude(_ra, unit=u.deg)
dec = Latitude(_dec, unit=u.deg)
cs = SkyCoord(ra, dec, frame='icrs')

loc = EarthLocation(lat=41.3 * u.deg, lon=-74 * u.deg, height=390 * u.m)

times = Time(54000. + _time_delta, format='mjd')
tframes = AltAz(obstime=times, location=loc)

altaz = cs.transform_to(tframes)  # 30+ seconds
```
The last function can take very long to run, and in fact this is a show-stopper for me to use astropy coordinates for such astrometric calculations.

## Accuracy
I made plots of the deviations. It seems, that with the proposed method (`_apci13_fast` with 10 min granularity) most points have less than 5 mas deviation from the original solution. No difference is apparent between `erfa.apci13` and the current implementation. The maximal deviation between `_apci13_fast` and current is of the order of 50 mas.

![az_diff](https://cloud.githubusercontent.com/assets/6582333/26033529/9a23886c-38ae-11e7-950c-62b623ac51f9.png)
![alt_diff](https://cloud.githubusercontent.com/assets/6582333/26033524/8f06514e-38ae-11e7-9652-54cb0f7291d4.png)

Zooming-in on the azimuthal deviations shows a pattern, which is introduced by the 10-min grid (0.007 d):
![az_diff_zoom](https://cloud.githubusercontent.com/assets/6582333/26033531/9f4966d6-38ae-11e7-9bbf-fc6cc5dc9ed9.png)

## Timing
There is a large performance bottleneck in the `cirs_to_icrs` and `icrs_to_cirs` transform functions. Both calculate the astrom struct using `erfa.apci`. The latter needs several parameters. One can already improve the speed somewhat, by using `erfa.apci13`, which only needs time as parameter. A much larger speed-up is obtained, if the time granularity is changed to 10 minutes or so (see above).

| Method| Runtime|
| :-------------| :-----|
| current |  35.2 s |
| `erfa.apci13` | 23.5 s |
| `_apci13_fast` | 295 ms (2.8 s, slowest loop)|

I'm not quite sure, why the (inital?) loop seems to be about 10x slower, when using `_apci13_fast`. Profiling suggests, that a lot of time is spent in `astropy._erfa._core._dtdb`, which is related to the line
```python
jd1, jd2 = get_jd12(cirs_frame.obstime, 'tdb')
```
According to the erfa docs, 'utc' should be as good as 'tdb' here, and doesn't seem to have the described issue. [update]I just realized, that for much finer resolution (e.g., 0.1 s), the difference between utc and tdb in fact becomes important. So for the moment, I switched back to the use of tdb.[/update]

## From proof-of-concept to a solution
While the quick-and-dirty solution already seems to provide a nice speed-up, there are other things to consider. The astrom struct is of course not only calculated in the `cirs_to_icrs` and `icrs_to_cirs` transform functions (although here, the performance hit is most critical). It would be much nicer, if it could be cached in a low-level routine. Unfortunately, since the erfa core is auto-generated (as a C-extension), this seems very difficult to achieve (at least for me). This would have been easier with the older Cython-based erfa wrapper. Likewise, a higher-level cache (in SkyCoord?) would be feasible, which stores already computed astrom values, but I guess, that changes would need to be made at many different places. Not sure, which option would be best.

Furthermore, the granularity (`MJD_RESOLUTION`) should be modifiable by the user, especially if higher accuracy is needed than provided by the value of 10 min.