@StuartLittlefair 2017-05-15
Hi @bwinkel,

First of all, thanks for submitting a PR. It's great to have people contributing, especially when they provide such detailed notes and justifications!

That said, I'm afraid I'm not a fan of this PR for a couple of reasons. 

The first is a philosophical position, that I think it's important that a hypothetical user of astropy can use the co-ordinate package without worrying about accuracy. Obviously there's a trade-off to be made between accuracy and speed but I think we need to think carefully before intentionally degrading the accuracy of an existing code. For example, someone currently using astropy to predict lunar occultations might get bitten by this.

The second is that the use of ```erfa.apci``` was introduced in #5273 to allow the use of (e.g) JPL ephemerides when performing coordinate transformations and light-travel-time corrections. I'd strongly argue there's a need to keep the GCRS-ICRS coordinate transform as-is (it's used in light-travel-time corrections), so your PR would make the ICRS-CIRS transform inconsistent with all the others, which I'm also not keen on. As a minimum, I'd want the proposed transform to be a non-default option.

Since your own use case seems rather extreme to drive the default behaviour in astropy, perhaps the best way forward  is to temporarily replace the transformations in astropy's frame transform graph *in your own code* - i.e if you want to use any particular function to transform CIRS->ICRS, just decorate that function with ```@frame_transform_graph.transform(FunctionTransform, CIRS, CIRS)``` and astropy will use your transformation and not the default.