@bwinkel 2017-05-15
Thanks a lot for the feedback, Stuart. I suspected that `erfa.apci` was there for the potential use of ephemerides. It'd probably be useful to add a comment to the source code. By default, astropy doesn't use to-date ephemerides though, which already leads to a certain loss in accuracy. I don't know how much this is; if I remember correctly the SOFA manual mentions microarcseconds somewhere. And you're absolutely right, the default behavior should be to have at least this kind of precision - either by not using caching at all by default, or by using a suitable `MJD_RESOLUTION`.

Note, that caching the `astrom` struct is not tied to the use of `acpi13`. One could easily adapt the function to work with the original `acpi`-based calculation.

I must also say that the use case I described is all but extreme. Almost all radio telescope, I know of, nowadays sample data with a resolution of ~1 s; for some types of observations, even faster (1 ms to 100 ms). Typical observing sessions are a couple of hours. At the moment, they just don't use astropy for these transformations, but I could imagine that some would, if possible.