@StuartLittlefair 2017-05-15
Thanks for the response. I still don't quite follow why sampling data at fast rates means that you have to be able to convert positions at 10**5 times in less than one second? Nevertheless, it is definitely true that for many people the astropy coordinate package is not as fast as it needs to be; I didn't mean to sound so negative above.

By describing your use case as extreme I simply mean that some users will demand high speed, and some will demand high accuracy and we need to balance the two (or find a way to satisfy both). I do think this PR could be a really valuable addition. However, given my concerns above, I think this PR should probably end up satisfying the following requirements:

1) Leave the existing behaviour as the default.
2) Add a way of modifying the behaviour, such as a ```ScienceState```, e.g:

```python
with approximate_transforms:
    fk5c.transform_to(GCRS(obstime=Time("J2000")))
```

3) Apply those changes to *all* relevant parts of the transform graph. One should get the same results if you go from ICRS->AltAz as ICRS->GCRS->AltAz, for example.
4) Carefully document the effect on precision in both position and time.

(2) and (3) could be achieved by writing a utility function to calculate the ```astrom``` context, that calls either the existing code, or your faster version depending on the results of the ```ScienceState``` object.