@bwinkel 2017-05-15
I like the idea of introducing a context manager for low-precision transforms. Will wait a little bit for other comments and then try to find out, how one could implement it. Thanks a lot!