@StuartLittlefair 2017-05-15
take a look at how it's done [here](https://github.com/astropy/astropy/blob/master/astropy/coordinates/solar_system.py). One thing that would need some thought is what would happen if a user both specified that the JPL ephemerides be used, and approximate transformations performed. Apart from giving that particular user a talking to, of course.