@bwinkel 2017-05-15
Again, as a proof of concept, I added a solution based on the ScienceState context. In the new `astrom_manager` module, there is a helper function `get_astrom` which encapsulates the MJD binning and allows to query the astrom structs. It has a `tcode` parameter, where one has to provide one of the erfa function names ('apio13', 'apci', 'apcs'). I didn't bother for now to setup a more fancy API for this, mainly because only three erfa (astrom) functions are used throughout all transformations.

It works like this:
```python
# see first comment for full code example
with transform_precision.set(precision):
    azels = cs.transform_to(tframes)
```
where `precision` is a float. If 0 (or less than 1e-3), no MJD binning is performed. Otherwise, precision is the bin width in seconds. 

I also realized, that the accuracy loss is much less than in the plots I made yesterday. Most of the high Azimuthal deviations where at the poles. I attach new plots, which show the angular-distance of the low-precision points to the original solution. For a binning of 600 s, the maximum error is 1.2 mas, for 10 s binning, I get 0.04 mas. 
![angsep10](https://cloud.githubusercontent.com/assets/6582333/26061837/5c5463b8-3989-11e7-87cc-a2d69281f720.png)
![angsep600](https://cloud.githubusercontent.com/assets/6582333/26061848/61f64d54-3989-11e7-8819-dcea561530eb.png)

I know, there is still a lot of work to do (tests, documentation, etc.). But I wanted to ask for your opinion first. Especially, if the design of the helper function is acceptable (I find it a bit clumsy, to be honest). Any ideas, how to improve?