@StuartLittlefair 2017-05-15
@bwinkel actually, I rather like the helper function to provide an astrometry context. It certainly cleans up the code a lot where it is actually used. 

I think that it might be a good idea to present the user with discrete choices for the precision (i.e a set of pre-chosen time resolutions) and document their accuracy. Otherwise we're giving end users work to do, working out if their chosen time resolution is precise enough for their needs. 

```python
with transform_precision.set('low'):
    azels = cs.transform_to(tframes)
```

I think at this point the general idea is firm enough for people like @eteq and @mhvk to look over and see if they approve in principle before putting in too much work with documentation/tests etc.