@bwinkel 2017-05-16
One thing that I don't like about the helper function is, that it is vital to plug the appropriate frame into it (in each transform function, two frames are involved, but only one of them must be used for the helper). Since the function is not exposed to the user, this may be acceptable, but it's not really nice from a design perspective.

I'm also thinking about how one could do some caching. If, for example, a conversion from AltAz to ICRS is performed, the MJD binning happens twice with the same obstimes. It would be cool, if the binned `astrom` values would be cached. At 10-min resolution, one would only need to store 720 instances to cover a full day. Furthermore, this would also speed-up the case where one repeatedly calls the transform functions with non-array frames (but also different obstimes). Does astropy have a cache utility class that could be used for such purposes?